Concentrate logic in OptimizedKeywordArguments
