+++ src/main/java/org/truffleruby/language/methods/Arity.java
@@ -152,10 +152,6 @@ public final class Arity {
         return keywordArguments;
     }
 
-    public boolean isKeywordArgumentOptimizable() {
-        return hasKeywords() && !hasKeywordsRest();
-    }
-
     public ArgumentDescriptor[] toAnonymousArgumentDescriptors() {
         List<ArgumentDescriptor> descs = new ArrayList<>();
